+++ Naotism/Source/Player.h
@@ -5,7 +5,7 @@ class Player : public Collidable{
 public:
 	Player::Player(float x = SCREENSIZES::LARGE.x / 2 , float y = SCREENSIZES::LARGE.y - 64 , sf::Vector2f velocity = sf::Vector2f(0 , 0) , int mass = 0);
 	Player::Player(const Player* /*&*/other);
-	virtual void Player::update(double delta);//override
+	virtual void Player::update(float delta);//override
 	virtual Player* Player::clone()const;
 	virtual void Player::collide( Collidable*& other);
 protected:
